\name{reedgraph-package}
\alias{reedgraph-package}
\alias{reedgraph}
\docType{package}
\title{
Graph routines for Maximum Commodity Flow and other miscellaneous
graph routines
}
\description{
Graph routines for Maximum Commodity Flow and other miscellaneous
graph routines. 
}
\details{
\tabular{ll}{
Package: \tab reedgraph\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-08-09\cr
License: \tab GPLv2 \cr
}

There are a number of utility functions supporting creating "random" networks
and max concurrent flow

}
\author{
Martin J. Reed

Maintainer: Martin J. Reed <martin@reednet.org.uk>
}
\references{
see Fleischer "Approximating fractional multicommodity flow independent of the number of
commodities", SIAM J. Discrete Maths, Vol 13/4, pp 505-520
}

\keyword{ graph }
\seealso{
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
